Remove IE-specific polyfill script tag
This PR removes the conditional comments and polyfill script tag that were previously used to support Internet Explorer (<!--[if IE]> ... <![endif]-->). The polyfill aimed to load features like Promise and ES6 capabilities for older IE versions. Given that support for IE has been deprecated and it is no longer a target browser, these lines are no longer necessary. This change simplifies the codebase and reduces dependency on external resources.